11/2/requirements.txt
cupy-cuda12x==13.3.0
fastrlock==0.8.2
numpy==2.2.0
